Deleted unwanted files and deleted 2 in file name in organism dir
