+++ API/index2.jsx
@@ -1,25 +1,24 @@
-import actionTypes from '../actionTypes/apiCalls.jsx';
-import Alert from '../components/organism/Alert';
+import actionTypes from "../actionTypes/apiCalls.jsx";
 
 const AdminApis = () => {
-    const { postActions, putActions, deleteActions } = actionTypes();
+  const { postActions, putActions, deleteActions } = actionTypes();
 
-    // create
-    const Posting = (endpoint, body) => {
-        postActions(endpoint, body);
-    };
+  // create
+  const Posting = (endpoint, body) => {
+    postActions(endpoint, body);
+  };
 
-    // update
-    const Putting = (endpoint, body) => {
-        putActions(endpoint, body);
-    };
+  // update
+  const Putting = (endpoint, body) => {
+    putActions(endpoint, body);
+  };
 
-    // delete
-    const Deleting = (endpoint, body) => {
-        deleteActions(endpoint, body);
-    };
+  // delete
+  const Deleting = (endpoint, body) => {
+    deleteActions(endpoint, body);
+  };
 
-    return { Posting, Putting, Deleting };
+  return { Posting, Putting, Deleting };
 };
 
 export default AdminApis;
