--- components/organism/AdminTabs/AllCategories.module.css
@@ -1,3 +1,0 @@
-.flex {
-    display: inline-block;
-}
