--- components/organism/AdminTabs/UpdateCategory.jsx
@@ -1,11 +1,0 @@
-import Form from "../Form2";
-
-const UpdateCategory = () => {
-  return (
-    <>
-      <Form />
-    </>
-  );
-};
-
-export default UpdateCategory;
