+++ components/organism/Handler.jsx
@@ -1,23 +1,26 @@
-import { useContext } from "react";
-
 import Api from "../../API";
-
+import { useContext } from "react";
 import { AdminContext } from "../../Context/AdminContext";
 
 const Handler = () => {
-  const { posting, deleteRoute } = Api();
+  const { posting, putting, deleteRoute } = Api();
 
-  const { setIsTab, setIsUpdatedTab, setContent, setTitle } =
+  const { setIsUpdatedTab, setIsTab, setIsForm, setImg } =
     useContext(AdminContext);
 
-  const handleUpdate = (id, title, description) => {
-    setIsTab("update");
+  const handleUpdateClick = (id, title, description, image) => {
     setIsUpdatedTab(true);
-    setTitle(title);
-    setContent(description);
+    setIsTab("update");
+    setIsForm({ id, title, description, image: "" });
+    setImg(image);
   };
 
-  const handleConfirm = (body) => {
+  const handleUpdate = (data) => {
+    putting("/api/category", data);
+  };
+
+  const handleConfirm = (id, title) => {
+    const body = { id, title };
     let answer = window.confirm("Are you sure you want to delete");
     if (answer) {
       handleDelete(body);
@@ -26,22 +29,23 @@ const Handler = () => {
 
   const handleDelete = (body) => deleteRoute("/api/category", body);
 
+  const handleCancel = () => {
+    setIsTab("home");
+  };
+
   const onSubmit = (data) => {
+    console.log("data", data);
     setIsTab("home");
     posting("/api/category", data);
   };
 
-  const onChangeImage = (imageList, setImages, setValue) => {
-    const { data_url } = imageList[0];
-    setImages(data_url);
-    setValue("image", data_url);
-  };
-
   return {
+    handleUpdateClick,
     handleUpdate,
     handleConfirm,
+    handleDelete,
+    handleCancel,
     onSubmit,
-    onChangeImage,
   };
 };
